Resetting handler for /monitored-endpoints ...
So that each test has a handler wired for that url and for each test to start over with an empty list of endpoints. An empty list of enpoints makes the NoData screen to get shown, forcing the tester to ensure to setup the list of endpoints for the context of the scenario.